Add native function to manually trigger gc
