+++ Cargo.toml
@@ -6,7 +6,7 @@ license = "Apache-2.0"
 authors = ["Fred Roth"]
 description = "An interpreter for the Lox language (http://craftinginterpreters.com/)"
 [features]
-default = ["stress_gc"]
+#default = ["stress_gc"]
 stress_gc = []
 
 [dependencies]
